requirements-faiss.txt
azureml-sdk==1.55.0
azure-keyvault==4.2.0
azure-monitor-opentelemetry==1.6.8
azure-storage-blob==12.19.1
bs4==0.0.2
faiss-cpu==1.7.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23    
langchain-openai==0.0.2
office365-rest-python-client==2.5.6  
openpyxl==3.1.2
pandas==2.2.0
pymsteams==0.2.2
unidecode==1.3.8